requirements.txt
audioread==3.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.0
cycler==0.12.1
decorator==5.2.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
grad-cam==1.5.5
huggingface-hub==0.30.2
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
msgpack==1.1.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
pooch==1.8.2
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.3
triton==3.2.0
ttach==0.0.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
zipp==3.21.0
